Added Lyapunov solvers.
I implemented a Lyapunov solver interface + another Memoizer for this. I also added three different implementations: a vectorized one, an iterative doubling one from Alex Maeder and the one that was already present in the other version of quatrex.

Everything works on batched inputs and is tested with some (stabilized) random inputs.