Put everything on GPU that can be put on GPU
I switched all the inversions to be GPU/CPU agnostic and made it a little bit more explicit when we actually need to call numpy.

I also added missing docstring and removed an unused method in the nevp.utils module.

Finally, i added tests for the memoizers and for the x_ii_formula kwarg in the Spectral solver.