First steps towards generalized, vectorized item-assignment in DSBSparse.
Similar to what i did for the _set_block/_get_block methods, the __setitem__/__getitem__ methods now wrap _set_items/_get_items. These functions can set a list of elements in a vectorized way. This also works on a specific part of the stack and both when distributed through stack and nnz.