Add OBCBlocks class and integrate into GFSolver implementations.
Now the obc blocks are passed directly into the Green's function solvers. With this we don't explicitely have to densify those blocks in the DSBSparse inputs. This is also implemented in the Inv solver.

In the RGF i added the option for also computing this Meir-Wingreen current throughout the device. I do this differently than in the old code, now i only compute everything during the second pass, no additional memory required but the quantities are kind of computed in both passes.

I also changed a couple things where we did not exploit the skew-hermitian symmetry correctly before. Fewer block-densifications now.