DOS peak filtering, perfectly periodic OBC blocks.
Since we noticed that the DOS peak filtering with these f1/f2 conditions does not actually do much, i added a new method solely based on limiting the DOS gradient for both G and W.

Second big change is that we now correctly construct the periodic blocks used to compute the OBC with the get_periodic_superblocks util method. This is used in the compute_obc methods of the subsystem solvers.

Last things are that the Hamitlonian is now not cut at all and that we make sure that all the input matrices are actually Hermitian.